requirements.txt
python-dotenv
streamlit
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-chroma
chromadb>=0.5.0
pypdf
docx2txt
tiktoken
sentence-transformers
python_version < "3.12"
streamlit>=1.29
chromadb==0.4.24
pysqlite3-binary==0.5.4

